cleanup: move layout helpers
