+++ src/layout/segmented.c
@@ -289,67 +289,3 @@ void reset_layout_segmented(segmented_layout_data* data) {
         reset_vgmstream(data->segments[i]);
     }
 }
-
-/* helper for easier creation of segments */
-VGMSTREAM* allocate_segmented_vgmstream(segmented_layout_data* data, int loop_flag, int loop_start_segment, int loop_end_segment) {
-    VGMSTREAM* vgmstream = NULL;
-    int channel_layout;
-    int i, sample_rate;
-    int32_t num_samples, loop_start, loop_end;
-    coding_t coding_type = data->segments[0]->coding_type;
-
-    /* save data */
-    channel_layout = data->segments[0]->channel_layout;
-    num_samples = 0;
-    loop_start = 0;
-    loop_end = 0;
-    sample_rate = 0;
-    for (i = 0; i < data->segment_count; i++) {
-        /* needs get_samples since element may use play settings */
-        int32_t segment_samples = vgmstream_get_samples(data->segments[i]);
-        int segment_rate = data->segments[i]->sample_rate;
-
-        if (loop_flag && i == loop_start_segment)
-            loop_start = num_samples;
-
-        num_samples += segment_samples;
-
-        if (loop_flag && i == loop_end_segment)
-            loop_end = num_samples;
-
-        /* inherit first segment's layout but only if all segments' layout match */
-        if (channel_layout != 0 && channel_layout != data->segments[i]->channel_layout)
-            channel_layout = 0;
-
-        if (sample_rate < segment_rate)
-            sample_rate = segment_rate;
-
-        if (coding_type == coding_SILENCE)
-            coding_type = data->segments[i]->coding_type;
-    }
-
-    /* respect loop_flag even when no loop_end found as it's possible file loops are set outside */
-
-
-    /* build the VGMSTREAM */
-    vgmstream = allocate_vgmstream(data->output_channels, loop_flag);
-    if (!vgmstream) goto fail;
-
-    vgmstream->meta_type = data->segments[0]->meta_type;
-    vgmstream->sample_rate = sample_rate;
-    vgmstream->num_samples = num_samples;
-    vgmstream->loop_start_sample = loop_start;
-    vgmstream->loop_end_sample = loop_end;
-    vgmstream->coding_type = coding_type;
-    vgmstream->channel_layout = channel_layout;
-
-    vgmstream->layout_type = layout_segmented;
-    vgmstream->layout_data = data;
-
-    return vgmstream;
-
-fail:
-    if (vgmstream) vgmstream->layout_data = NULL;
-    close_vgmstream(vgmstream);
-    return NULL;
-}
